Clean up unused codes and group common code in reusable methods of the DocumentRepository class
