+++ src/main/java/org/silverpeas/migration/classified/XmlFormImagesToAttachment.java
@@ -30,40 +30,48 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.Collection;
-
+import java.util.Locale;
+import javax.jcr.RepositoryException;
+import javax.jcr.Session;
 import org.silverpeas.dbbuilder.dbbuilder_dl.DbBuilderDynamicPart;
 import org.silverpeas.migration.jcr.service.AttachmentException;
-import org.silverpeas.migration.jcr.service.SimpleDocumentService;
+import org.silverpeas.migration.jcr.service.RepositoryManager;
 import org.silverpeas.migration.jcr.service.model.DocumentType;
+import org.silverpeas.migration.jcr.service.model.SimpleDocument;
 import org.silverpeas.migration.jcr.service.model.SimpleDocumentPK;
+import org.silverpeas.migration.jcr.service.repository.DocumentRepository;
+import org.silverpeas.util.Console;
 import org.silverpeas.util.StringUtil;
-
 
 /**
  * @author cbonin
  */
 public class XmlFormImagesToAttachment extends DbBuilderDynamicPart {
-  
-  private final SimpleDocumentService service;
-  
+
+  private final RepositoryManager repositoryManager;
+  private final DocumentRepository documentRepository;
+
   public XmlFormImagesToAttachment() {
-    this.service = new SimpleDocumentService();
+    this.repositoryManager = new RepositoryManager();
+    this.documentRepository = new DocumentRepository(repositoryManager);
   }
 
   public void migrate() throws IOException, SQLException {
+    Session session = openJCRSession();
     try {
-
       getConsole()
-          .printMessage("Migrate Classified Images in XML Form To Attachment and Classified Description in XML Form to Classified DB");
+          .printMessage(
+          "Migrate Classified Images in XML Form To Attachment and Classified Description in XML Form to Classified DB");
       getConsole().printMessage("");
+      long migrationStart = System.currentTimeMillis();
 
       Collection<String> listInstanceId = getListClassifiedInstance();
-
       for (String instanceId : listInstanceId) {
+        long instanceMigrationStart = System.currentTimeMillis();
         int instanceIntId = Integer.parseInt(instanceId);
         getConsole().printMessage("Migrate Classified instance id = " + instanceId);
         String xmlFormName = getXMLFormName(instanceIntId);
-        getConsole().printTrace("Migrate Classified XML Form = "+xmlFormName);
+        getConsole().printTrace("Migrate Classified XML Form = " + xmlFormName);
         Collection<Integer> listTemplateId = getListTemplate(xmlFormName, instanceId);
         for (Integer templateId : listTemplateId) {
           getConsole().printTrace("TemplateId = " + templateId);
@@ -72,76 +80,74 @@ public class XmlFormImagesToAttachment extends DbBuilderDynamicPart {
             int recordId = record.getRecordId();
             String classifiedId = record.getExternalId();
             int classifiedIntId = Integer.parseInt(classifiedId);
-            getConsole().printMessage("Record [recordId = " + recordId + ", externalId = " +
-                classifiedId + "]");
+            getConsole().printMessage("Record [recordId = " + recordId + ", externalId = "
+                + classifiedId + "]");
             Collection<FieldTemplate> listValue = getListValue(recordId);
             for (FieldTemplate fieldTemplate : listValue) {
-              getConsole().printTrace("Field Name = "+fieldTemplate.getFieldName());//category | type | description | photo 
+              getConsole().printTrace("Field Name = " + fieldTemplate.getFieldName());//category | type | description | photo
               if (fieldTemplate.getFieldName().startsWith("photo")) {
                 if (StringUtil.isDefined(fieldTemplate.getFieldValue())) {
                   getConsole().printTrace("Photo field value = " + fieldTemplate.getFieldValue());
-                  
-                  getConsole().printTrace("Delete field template recordId = " + recordId +
-                      ", fieldName = '" + fieldTemplate.getFieldName() +
-                      "', fieldValue = " + fieldTemplate.getFieldValue());
+
+                  getConsole().printTrace("Delete field template recordId = " + recordId
+                      + ", fieldName = '" + fieldTemplate.getFieldName() + "', fieldValue = "
+                      + fieldTemplate.getFieldValue());
                   try {
                     deletePhotoValue(recordId, fieldTemplate.getFieldName());
                   } catch (SQLException e) {
                     getConsole()
-                        .printError("ERROR when Deleting field template recordId = " +
-                        recordId + ", fieldName = '" + fieldTemplate.getFieldName() +
-                        "', fieldValue = " + fieldTemplate.getFieldValue() +
-                        ", error = " + e.getMessage(), e);
+                        .printError("ERROR when Deleting field template recordId = " + recordId
+                        + ", fieldName = '" + fieldTemplate.getFieldName() + "', fieldValue = "
+                        + fieldTemplate.getFieldValue() + ", error = " + e.getMessage(), e);
                   }
 
-                  if(StringUtil.isLong(fieldTemplate.getFieldValue())) {
-                    SimpleDocumentPK simpleDocumentPk = new SimpleDocumentPK(null, "classifieds"+instanceId);
+                  if (StringUtil.isLong(fieldTemplate.getFieldValue())) {
+                    SimpleDocumentPK simpleDocumentPk = new SimpleDocumentPK(null, "classifieds"
+                        + instanceId);
                     simpleDocumentPk.setOldSilverpeasId(Long.valueOf(fieldTemplate.getFieldValue()));
                     try {
-                      this.service.moveImageContext(simpleDocumentPk, getConsole());
+                      moveImageContext(session, simpleDocumentPk);
                     } catch (AttachmentException e) {
                       getConsole()
-                      .printError("ERROR when Moving attachment image id = " +fieldTemplate.getFieldValue()+
-                                  ", "+
-                                  e.getMessage());
+                          .printError("ERROR when Moving attachment image id = " + fieldTemplate.
+                          getFieldValue() + ", " + e.getMessage());
                     }
-                  } 
+                  }
                 } else {
                   getConsole().printTrace("Photo field value = null");
-                  getConsole().printTrace("Delete field template recordId = " + recordId +
-                      ", fieldName = '" + fieldTemplate.getFieldName() + "', fieldValue = null");
+                  getConsole().printTrace("Delete field template recordId = " + recordId
+                      + ", fieldName = '" + fieldTemplate.getFieldName() + "', fieldValue = null");
                   try {
                     deletePhotoValue(recordId, fieldTemplate.getFieldName());
                   } catch (SQLException e) {
-                    getConsole().printError("ERROR when Deleting field template recordId = " +
-                        recordId + ", fieldName = '" + fieldTemplate.getFieldName() +
-                        "', fieldValue = null, error = " +
-                        e.getMessage(), e);
+                    getConsole().printError("ERROR when Deleting field template recordId = "
+                        + recordId + ", fieldName = '" + fieldTemplate.getFieldName()
+                        + "', fieldValue = null, error = " + e.getMessage(), e);
                   }
                 }
               } else if ("description".equals(fieldTemplate.getFieldName())) {
                 getConsole()
                     .printTrace("Description field value = " + fieldTemplate.getFieldValue());
-                getConsole().printTrace("Update classified instanceId = " + instanceId +
-                    ", classifiedId = " + classifiedId + ", description = " +
-                    fieldTemplate.getFieldValue());
+                getConsole().printTrace("Update classified instanceId = " + instanceId
+                    + ", classifiedId = " + classifiedId + ", description = " + fieldTemplate.
+                    getFieldValue());
                 boolean updated = false;
                 try {
                   updateClassified(instanceId, classifiedIntId, fieldTemplate.getFieldValue());
                   updated = true;
                 } catch (SQLException e) {
-                  getConsole().printError("ERROR when Updating classified instanceId = " +
-                      instanceId + ", classifiedId = " + classifiedId + ", description = " +
-                      fieldTemplate.getFieldValue() + ", error = " + e.getMessage(), e);
+                  getConsole().printError("ERROR when Updating classified instanceId = "
+                      + instanceId + ", classifiedId = " + classifiedId + ", description = "
+                      + fieldTemplate.getFieldValue() + ", error = " + e.getMessage(), e);
                 }
                 if (updated) {
-                  getConsole().printTrace("Delete field template recordId = " + recordId +
-                      ", fieldName = 'description'");
+                  getConsole().printTrace("Delete field template recordId = " + recordId
+                      + ", fieldName = 'description'");
                   try {
                     deleteDescriptionValue(recordId);
                   } catch (SQLException e) {
-                    getConsole().printError("ERROR when Deleting field template recordId = " +
-                        recordId + ", fieldName = 'description', error = " + e.getMessage(), e);
+                    getConsole().printError("ERROR when Deleting field template recordId = "
+                        + recordId + ", fieldName = 'description', error = " + e.getMessage(), e);
                   }
                 }
               }
@@ -151,10 +157,18 @@ public class XmlFormImagesToAttachment extends DbBuilderDynamicPart {
           getConsole().printTrace("---------------------");
         }
 
+        long instanceMigrationEnd = System.currentTimeMillis();
+        getConsole().printMessage("Classified instance " + instanceId + " has been migrated in "
+            + (instanceMigrationEnd - instanceMigrationStart) + " ms");
         getConsole()
             .printMessage("--------------------------------------------------------------------");
       }
+      long migrationEnd = System.currentTimeMillis();
+      getConsole().printMessage("Classified components have been migrated in "
+          + (migrationEnd - migrationStart) + " ms");
     } finally {
+      repositoryManager.logout(session);
+      repositoryManager.shutdown();
       getConsole().close();
     }
   }
@@ -215,8 +229,8 @@ public class XmlFormImagesToAttachment extends DbBuilderDynamicPart {
     try {
       pstmt =
           getConnection()
-              .prepareStatement(
-                  "SELECT templateId FROM sb_formtemplate_template where templateName=? and externalId like ?");
+          .prepareStatement(
+          "SELECT templateId FROM sb_formtemplate_template where templateName=? and externalId like ?");
       pstmt.setString(1, xmlFormName);
       pstmt.setString(2, "classifieds" + instanceId + "%");
       rs = pstmt.executeQuery();
@@ -298,8 +312,8 @@ public class XmlFormImagesToAttachment extends DbBuilderDynamicPart {
     try {
       pstmt =
           getConnection()
-              .prepareStatement(
-                  "UPDATE SC_Classifieds_Classifieds set description = ? where instanceId = ? and classifiedId = ? ");
+          .prepareStatement(
+          "UPDATE SC_Classifieds_Classifieds set description = ? where instanceId = ? and classifiedId = ? ");
       pstmt.setString(1, description);
       pstmt.setString(2, "classifieds" + instanceId);
       pstmt.setInt(3, classifiedId);
@@ -338,5 +352,66 @@ public class XmlFormImagesToAttachment extends DbBuilderDynamicPart {
         pstmt.close();
       }
     }
+  }
+
+  private void moveImageContext(Session session, SimpleDocumentPK simpleDocumentPk) {
+    try {
+      Console console = getConsole();
+      session = repositoryManager.getSession();
+      SimpleDocument simpleDocument = documentRepository.findDocumentByOldSilverpeasId(session,
+          simpleDocumentPk.getComponentName(),
+          simpleDocumentPk.getOldSilverpeasId(), false, null);
+      if (simpleDocument != null) {
+        boolean verifFormatImage = verifFormatImage(simpleDocument.getFilename());
+        if (verifFormatImage) {
+          console.printTrace("Delete attachment with attachmentId = " + simpleDocument.getId()
+              + ", oldSilverpeasId = " + simpleDocument.getOldSilverpeasId());
+          String sourcePath = simpleDocument.getFullJcrPath();
+          simpleDocument.setDocumentType(DocumentType.attachment);
+          String destinationPath = simpleDocument.getFullJcrPath();
+          console.printTrace("Moving " + sourcePath + " to " + destinationPath);
+          if (!session.nodeExists(destinationPath)) {
+            documentRepository.
+                prepareComponentAttachments(session, simpleDocumentPk.getInstanceId(),
+                DocumentType.attachment.getFolderName());
+          }
+          session.move(sourcePath, destinationPath);
+        } else {// format Image not correct
+          console.printTrace("Format Image not correct, delete attachment attachmentId = "
+              + simpleDocument.getId() + ", oldSilverpeasId = " + simpleDocument
+              .getOldSilverpeasId());
+          documentRepository.deleteDocument(session, simpleDocument.getPk());
+        }
+        session.save();
+      } else {
+        throw new AttachmentException("ERROR Simple Document with oldSilverpeasId "
+            + simpleDocumentPk.getOldSilverpeasId() + " not found");
+      }
+    } catch (RepositoryException ex) {
+      throw new AttachmentException(ex);
+    }
+  }
+
+  private Session openJCRSession() {
+    try {
+      return repositoryManager.getSession();
+    } catch (RepositoryException ex) {
+      throw new AttachmentException(ex);
+    }
+  }
+
+  private boolean verifFormatImage(String filename) {
+    int indexPoint = filename.lastIndexOf('.');
+    if (indexPoint != -1) {
+      // le fichier contient une extension. On recupere l'extension
+      String extension = filename.substring(indexPoint + 1);
+      extension = extension.toLowerCase(Locale.FRANCE);
+      if ("jpg".equals(extension) || "gif".equals(extension) || "bmp".equals(extension) || "tiff"
+          .equals(extension) || "tif".equals(extension) || "jpeg".equals(extension) || "png".equals(
+          extension)) {
+        return true;
+      }
+    }
+    return false;
   }
 }
