+++ src/main/java/org/silverpeas/migration/jcr/service/AttachmentService.java
@@ -33,7 +33,6 @@ import org.silverpeas.migration.jcr.service.model.SimpleDocument;
 import org.silverpeas.migration.jcr.service.model.SimpleDocumentPK;
 import org.silverpeas.migration.jcr.service.model.UnlockContext;
 import org.silverpeas.migration.jcr.service.model.WAPrimaryKey;
-import org.silverpeas.util.Console;
 
 /**
  * @author ehugonnet
@@ -169,14 +168,6 @@ public interface AttachmentService {
    * @param document the document to deleted.
    */
   void deleteAttachment(SimpleDocument document);
-
-  /**
-   * Find document image, and move to Attachment context
-   *
-   * @param primaryKey the primary key of the document.
-   * @param console to print message.
-   */
-  public void moveImageContext(SimpleDocumentPK primaryKey, Console console);
 
   /**
    * Find all components that contains wysiwyg.
