+++ src/main/java/org/silverpeas/migration/jcr/service/SimpleDocumentService.java
@@ -30,23 +30,16 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
-import java.util.Locale;
 import javax.jcr.RepositoryException;
 import javax.jcr.Session;
 import org.apache.commons.io.IOUtils;
-import org.silverpeas.migration.jcr.service.model.DocumentType;
-import org.silverpeas.migration.jcr.service.model.ForeignPK;
 import org.silverpeas.migration.jcr.service.model.SimpleDocument;
 import org.silverpeas.migration.jcr.service.model.SimpleDocumentPK;
-import org.silverpeas.migration.jcr.service.model.UnlockContext;
-import org.silverpeas.migration.jcr.service.model.WAPrimaryKey;
 import org.silverpeas.migration.jcr.service.repository.DocumentRepository;
-import org.silverpeas.util.Console;
 import org.silverpeas.util.StringUtil;
 
-public class SimpleDocumentService implements AttachmentService {
+public class SimpleDocumentService {
 
   private final DocumentRepository repository;
   private final RepositoryManager repositoryManager;
@@ -62,7 +55,7 @@ public class SimpleDocumentService implements AttachmentService {
     repository = new DocumentRepository(repositoryManager);
   }
 
-  public DocumentRepository getRepository() {
+  public DocumentRepository getDocumentRepository() {
     return repository;
   }
 
@@ -70,37 +63,6 @@ public class SimpleDocumentService implements AttachmentService {
     return repositoryManager;
   }
 
-  /**
-   * Create file attached to an object who is identified by the foreignId.
-   *
-   * @param document the document to be created.
-   * @param content the binary content of the document.
-   * @return the stored document.
-   */
-  @Override
-  public SimpleDocument createAttachment(SimpleDocument document, InputStream content)
-      throws AttachmentException {
-    Session session = null;
-    try {
-      session = repositoryManager.getSession();
-      SimpleDocumentPK docPk = repository.createDocument(session, document);
-      session.save();
-      SimpleDocument createdDocument = repository.findDocumentById(session, docPk, document.
-          getLanguage());
-      createdDocument.setPublicDocument(document.isPublic());
-      SimpleDocument finalDocument = repository.unlock(session, createdDocument, false);
-      repository.storeContent(finalDocument, content);
-      return finalDocument;
-    } catch (RepositoryException ex) {
-      throw new AttachmentException(ex);
-    } catch (IOException ex) {
-      throw new AttachmentException(ex);
-    } finally {
-      repositoryManager.logout(session);
-    }
-  }
-
-  @Override
   public SimpleDocument searchDocumentById(SimpleDocumentPK primaryKey, String lang) {
     Session session = null;
     try {
@@ -123,55 +85,6 @@ public class SimpleDocumentService implements AttachmentService {
     }
   }
 
-  @Override
-  public List<SimpleDocument> listAllDocumentsByForeignKey(WAPrimaryKey foreignKey, String lang) {
-    Session session = null;
-    try {
-      session = repositoryManager.getSession();
-      return repository.listAllDocumentsByForeignId(session, foreignKey.getInstanceId(), foreignKey
-          .
-          getId(), lang);
-    } catch (RepositoryException ex) {
-      throw new AttachmentException(ex);
-    } finally {
-      repositoryManager.logout(session);
-    }
-  }
-
-  @Override
-  public List<SimpleDocument> listDocumentsByForeignKey(WAPrimaryKey foreignKey, String lang) {
-    Session session = null;
-    try {
-      session = repositoryManager.getSession();
-      return repository.listDocumentsByForeignId(session, foreignKey.getInstanceId(), foreignKey.
-          getId(), lang);
-    } catch (RepositoryException ex) {
-      throw new AttachmentException(ex);
-    } finally {
-      repositoryManager.logout(session);
-    }
-  }
-
-  @Override
-  public void updateAttachment(SimpleDocument document) throws AttachmentException {
-    Session session = null;
-    try {
-      session = repositoryManager.getSession();
-      SimpleDocument oldAttachment = repository.findDocumentById(session, document.getPk(), document
-          .getLanguage());
-      repository.fillNodeName(session, document);
-      repository.updateDocument(session, document);
-      session.save();
-    } catch (RepositoryException ex) {
-      throw new AttachmentException(ex);
-    } catch (IOException ex) {
-      throw new AttachmentException(ex);
-    } finally {
-      repositoryManager.logout(session);
-    }
-  }
-
-  @Override
   public void updateAttachment(SimpleDocument document, File content) {
     InputStream in = null;
     try {
@@ -184,7 +97,6 @@ public class SimpleDocumentService implements AttachmentService {
     }
   }
 
-  @Override
   public void updateAttachment(SimpleDocument document, InputStream in) {
     Session session = null;
     try {
@@ -213,130 +125,6 @@ public class SimpleDocumentService implements AttachmentService {
     }
   }
 
-  @Override
-  public SimpleDocument createAttachment(SimpleDocument document, File content) {
-    InputStream in = null;
-    try {
-      in = new BufferedInputStream(new FileInputStream(content));
-      return createAttachment(document, in);
-    } catch (FileNotFoundException ex) {
-      throw new AttachmentException(ex);
-    } finally {
-      IOUtils.closeQuietly(in);
-    }
-  }
-
-  /**
-   * Release a locked file.
-   *
-   * @param context : the unlock parameters.
-   * @return false if the file is locked - true if the unlock succeeded.
-   * @throws AttachmentException
-   */
-  @Override
-  public boolean unlock(UnlockContext context) {
-    Session session = null;
-    try {
-      session = repositoryManager.getSession();
-      SimpleDocument document = repository.findDocumentById(session,
-          new SimpleDocumentPK(context.getAttachmentId()), context.getLang());
-      if (!context.isForce() && document.isReadOnly() && !document.getEditedBy().equals(context.
-          getUserId())) {
-        return false;
-      }
-      if (context.isWebdav() || context.isUpload()) {
-        String workerId = document.getEditedBy();
-        document.setUpdated(new Date());
-        document.setUpdatedBy(workerId);
-      }
-      document.setPublicDocument(context.isPublicVersion());
-      document.setComment(context.getComment());
-      SimpleDocument finalDocument = repository.unlock(session, document, context.isForce());
-
-      File file = new File(finalDocument.getAttachmentPath());
-      if (!file.exists() && !context.isForce()) {
-        repository.duplicateContent(document, finalDocument);
-      }
-      session.save();
-    } catch (IOException e) {
-      throw new AttachmentException(e);
-    } catch (RepositoryException e) {
-      throw new AttachmentException(e);
-    } finally {
-      repositoryManager.logout(session);
-    }
-    return true;
-  }
-
-  /**
-   * Lock a file so it can be edited by an user.
-   *
-   * @param attachmentId
-   * @param userId
-   * @param language
-   * @return false if the attachment is already checkout - true if the attachment was successfully
-   * checked out.
-   */
-  @Override
-  public boolean lock(String attachmentId, String userId, String language) {
-    Session session = null;
-    try {
-      SimpleDocumentPK pk = new SimpleDocumentPK(attachmentId);
-      session = repositoryManager.getSession();
-      SimpleDocument document = repository.findDocumentById(session, pk, language);
-      if (document.isReadOnly()) {
-        return document.getEditedBy().equals(userId);
-      }
-      repository.lock(session, document, document.getEditedBy());
-      document.edit(userId);
-      updateAttachment(session, document);
-      session.save();
-      return true;
-    } catch (RepositoryException ex) {
-      throw new AttachmentException(ex);
-    } catch (IOException ex) {
-      throw new AttachmentException(ex);
-    } finally {
-      repositoryManager.logout(session);
-    }
-  }
-
-  private void updateAttachment(Session session, SimpleDocument document)
-      throws RepositoryException, IOException {
-    repository.updateDocument(session, document);
-  }
-
-  @Override
-  public SimpleDocument findExistingDocument(SimpleDocumentPK pk, String fileName,
-      ForeignPK foreign, String lang) {
-    List<SimpleDocument> exisitingsDocuments = listDocumentsByForeignKey(foreign, lang);
-    SimpleDocument document = searchDocumentById(pk, lang);
-    if (document == null) {
-      for (SimpleDocument doc : exisitingsDocuments) {
-        if (doc.getFilename().equalsIgnoreCase(fileName)) {
-          return doc;
-        }
-      }
-    }
-    return document;
-  }
-
-  @Override
-  public List<SimpleDocument> listDocumentsByForeignKeyAndType(WAPrimaryKey foreignKey,
-      DocumentType type, String lang) {
-    Session session = null;
-    try {
-      session = repositoryManager.getSession();
-      return repository.listDocumentsByForeignIdAndType(session, foreignKey.getInstanceId(),
-          foreignKey.getId(), type, lang);
-    } catch (RepositoryException ex) {
-      throw new AttachmentException(ex);
-    } finally {
-      repositoryManager.logout(session);
-    }
-  }
-
-  @Override
   public SimpleDocument mergeDocument(SimpleDocument document, SimpleDocument merge) {
     File content = new File(merge.getAttachmentPath());
     if (!content.exists() || !content.isFile()) {
@@ -350,7 +138,6 @@ public class SimpleDocumentService implements AttachmentService {
     return searchDocumentById(document.getPk(), document.getLanguage());
   }
 
-  @Override
   public List<String> listBasenames(String instanceId) {
     Session session = null;
     try {
@@ -363,7 +150,6 @@ public class SimpleDocumentService implements AttachmentService {
     }
   }
 
-  @Override
   public List<SimpleDocument> listWysiwygForBasename(String basename, String instanceId) {
     Session session = null;
     try {
@@ -376,7 +162,6 @@ public class SimpleDocumentService implements AttachmentService {
     }
   }
 
-  @Override
   public List<String> listComponentIdsWithWysiwyg() {
     Session session = null;
     try {
@@ -389,7 +174,6 @@ public class SimpleDocumentService implements AttachmentService {
     }
   }
 
-  @Override
   public void shutdown() {
     this.repositoryManager.shutdown();
   }
@@ -399,7 +183,6 @@ public class SimpleDocumentService implements AttachmentService {
    *
    * @param document the attachmentDetail object to deleted.
    */
-  @Override
   public void deleteAttachment(SimpleDocument document) {
     Session session = null;
     try {
@@ -407,61 +190,6 @@ public class SimpleDocumentService implements AttachmentService {
       repository.fillNodeName(session, document);
       repository.deleteDocument(session, document.getPk());
       session.save();
-    } catch (RepositoryException ex) {
-      throw new AttachmentException(ex);
-    } finally {
-      repositoryManager.logout(session);
-    }
-  }
-
-  private boolean verifFormatImage(String filename) {
-    int indexPoint = filename.lastIndexOf('.');
-    if (indexPoint != -1) {
-      // le fichier contient une extension. On recupere l'extension
-      String extension = filename.substring(indexPoint + 1);
-      extension = extension.toLowerCase(Locale.FRANCE);
-      if ("jpg".equals(extension) || "gif".equals(extension) || "bmp".equals(extension) || "tiff"
-          .equals(extension) || "tif".equals(extension) || "jpeg".equals(extension) || "png".equals(
-          extension)) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  @Override
-  public void moveImageContext(SimpleDocumentPK primaryKey, Console console) {
-    Session session = null;
-    try {
-      session = repositoryManager.getSession();
-      SimpleDocument simpleDocument = repository
-          .findDocumentByOldSilverpeasId(session, primaryKey.getComponentName(),
-          primaryKey.getOldSilverpeasId(), false, null);
-      if (simpleDocument != null) {
-        boolean verifFormatImage = verifFormatImage(simpleDocument.getFilename());
-        if (verifFormatImage) {
-          console.printTrace("Delete attachment with attachmentId = " + simpleDocument.getId()
-              + ", oldSilverpeasId = " + simpleDocument.getOldSilverpeasId());
-          String sourcePath = simpleDocument.getFullJcrPath();
-          simpleDocument.setDocumentType(DocumentType.attachment);
-          String destinationPath = simpleDocument.getFullJcrPath();
-          console.printTrace("Moving " + sourcePath + " to " + destinationPath);
-          if (!session.nodeExists(destinationPath)) {
-            repository.prepareComponentAttachments(primaryKey.getInstanceId(),
-                DocumentType.attachment.getFolderName());
-          }
-          session.move(sourcePath, destinationPath);
-        } else {// format Image not correct
-          console.printTrace("Format Image not correct, delete attachment attachmentId = "
-              + simpleDocument.getId() + ", oldSilverpeasId = " + simpleDocument
-              .getOldSilverpeasId());
-          repository.deleteDocument(session, simpleDocument.getPk());
-        }
-        session.save();
-      } else {
-        throw new AttachmentException("ERROR Simple Document with oldSilverpeasId " + primaryKey
-            .getOldSilverpeasId() + " not found");
-      }
     } catch (RepositoryException ex) {
       throw new AttachmentException(ex);
     } finally {
